Rename `reducers.js` to follow the proper naming scheme (not plural)
