+++ client/state/signup/reducer.js
@@ -6,7 +6,7 @@ import { combineReducers } from 'redux';
 /**
  * Internal dependencies
  */
-import dependencyStore from './dependency-store/reducers';
+import dependencyStore from './dependency-store/reducer';
 
 export default combineReducers( {
 	dependencyStore,
